Вынес заголовок с табами в отдельный компонент и перенес в BurgerIngredients
