+++ src/components/burger-constructor/burger-constructor.module.css
@@ -1,10 +1,11 @@
 .section {
   display: flex;
-  overflow: hidden;
-  flex-wrap: wrap;
+  flex-direction: column;
   align-items: flex-start;
   justify-content: flex-start;
   width: var(--main-panel-width);
+  height: 100%;
+  padding-top: 100px;
 
   --element-height: 80px;
 }
