+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,17 +1,25 @@
 .section {
   display: flex;
-  overflow: auto;
-  flex-wrap: wrap;
-  align-items: center;
+  flex-direction: column;
+  align-items: flex-start;
   justify-content: left;
   width: var(--main-panel-width);
+  height: 100%;
+}
+
+.types {
+  display: flex;
+  flex-direction: column;
+  overflow: auto;
+  width: 100%;
+  height: 100%;
   composes: scroll from '/src/index.css';
 }
 
-.section > div {
+.types > div {
   width: 100%;
 }
 
-.section > :first-child {
+.types > :first-child {
   margin-top: 0;
 }
